glovebox-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
google-cloud-firestore==2.19.0
google-auth==2.37.0
python-dotenv==1.0.1
requests==2.32.3
